backend/ml/prediction-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
mlflow==2.8.1
